create ufc module
